Allow for external definition of data centers.
The original version hard-coded the data centers and their servers figuring
that it would be valid for a while. But before this was officially released,
a new data center was created and the server assignments we changed. While
updating the list, it was moved out into a TOML file. The builtin list is
in a resource, but we try to load them from .chatter14.datacenters.toml or
chatter14.datacenters.toml that the user can create if the lists change before
this application is updated.

Abstracted out reading and parsing of the TOML files into a new helper.

Created a new Universe object to hold the data center definitions which is
populated from a TOML file.

Changed how errors are handled in the TOML files. All checking except for
TOML syntax errors are now handled in the `validate()` method. All errors
are listed instead of just the first one.

Updated 4koma to the latest version.