Added support for two types of tell.
It turns out that FFXIV has 2 different codes for tell: one for a tell sent by
me and one for a tell sent to me. This adds support for both types. In
addition, the name of the tell is the same in both cases: the other guy. So
we now rewrite the sender for tells from me to be my name. To support this
a new configuration option "me" has been added to allow the user to specify
their own name. If no name is given, we use "Me" as the name.